requirements.txt
winput
XInput-Python
